Extracted utility functions into a separate file
Also have been getting a red line under react-native-calendars import in
calendar screen so i tried to add it again using yarn but it didnt
remove the red line. Functionality of the calendar is not affected
though so i just left it
