+++ package.json
@@ -32,7 +32,7 @@
     "react-dom": "16.13.1",
     "react-hook-form": "^7.9.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-41.0.0.tar.gz",
-    "react-native-calendars": "^1.1263.0",
+    "react-native-calendars": "^1.1264.0",
     "react-native-gesture-handler": "~1.10.2",
     "react-native-get-random-values": "~1.7.0",
     "react-native-modal": "^11.10.0",
