+++ src/screens/home/CalendarScreen.tsx
@@ -3,10 +3,10 @@ import { useSelector } from 'react-redux';
 import { selectCards } from '../../redux/selectors/cards';
 import { View, Text, FlatList } from 'react-native';
 import { Calendar } from 'react-native-calendars';
-// import { DashboardCard } from '../../components/home/DashboardCard';
 import CalendarCard from '../../components/home/CalendarCard';
 import { CardModelWithUid } from '../../database/models/cards';
 import theme from '../../styles/theme.style';
+import { getDayText } from '../../util/dateFormat';
 
 const CalendarScreen: React.FC = ({navigation}: any) => {
 
@@ -63,27 +63,6 @@ const CalendarScreen: React.FC = ({navigation}: any) => {
 		return card.startdate.toDateString() === date.toDateString() && card.duedate.toDateString() === date.toDateString();
 	};
 	
-	const getDayText = (num: number): string => {
-		switch (num) {
-		case 1:
-			return 'Mon';
-		case 2:
-			return 'Tue';
-		case 3:
-			return 'Wed';
-		case 4:
-			return 'Thu';
-		case 5:
-			return 'Fri';
-		case 6:
-			return 'Sat';
-		case 0:
-			return 'Sun';
-		default:
-			return '';
-		}
-	};
-
 	const clickHandler = (card: CardModelWithUid) => navigation.navigate('UpdateCardModal', card);
 
 	const renderItem = (item: any) => <CalendarCard info={item} clickHandler={clickHandler}/>;
